requirements.txt
astroid==1.5.3
autopep8==1.3.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.6
django-debug-toolbar==1.8
django-modelcluster==3.1
django-redis==4.8.0
django-taggit==0.22.1
django-treebeard==4.1.2
django-username-email==2.1.2
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyaml==17.10.0
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.7.4
PyPDF2==1.26.0
pytz==2017.2
PyYAML==3.12
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wagtail==1.13
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.0
wrapt==1.10.11
